requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
apispec==6.0.2
attrs==23.2.0
bidict==0.22.1
click==8.1.3
dnspython==2.5.0
dominate==2.7.0
ecdsa==0.18.0
email-validator==2.1.0.post1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
flask-http-middleware==0.4.2
Flask-JWT-Extended==4.6.0
Flask-Moment==1.0.5
flask-restplus==0.13.0
flask-smorest==0.40.0
Flask-SocketIO==5.3.6
flask-toolkits==0.7.10
Flask-WTF==1.1.1
gunicorn==20.1.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.2
marshmallow==3.19.0
mongoengine==0.27.0
packaging==23.0
passlib==1.7.4
platformdirs==4.2.0
pyasn1==0.5.1
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==0.21.1
python-engineio==4.9.0
python-jose==3.3.0
python-socketio==5.11.1
pytz==2023.4
referencing==0.33.0
rpds-py==0.17.1
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.9.0
visitor==0.1.3
webargs==8.2.0
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1
yapf==0.40.2
zipp==3.17.0
